backend/requirements.txt
# FastAPI Backend - Python Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Task Queue (optional)
celery==5.3.6
flower==2.0.1

# Storage (S3)
boto3==1.34.34
botocore==1.34.34

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4
reportlab==4.0.9
weasyprint==60.2
pdfminer.six>=20221105    # Digital PDF text extraction
pdfrw>=0.4                # AcroForm read/write
pikepdf>=8.0.0            # Alternative PDF manipulation
jinja2>=3.1.2             # Template engine

# OCR (optional)
pytesseract==0.3.10
pillow==10.2.0

# AI/LLM
openai==1.10.0
anthropic==0.8.1
langchain==0.1.4

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Stripe
stripe==8.1.0

# Monitoring & Logging
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
sentry-sdk[fastapi]==1.40.0

# Validation & Utils
email-validator==2.1.0
python-slugify==8.0.1
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.14
mypy==1.8.0
